Start Python code to generate large matrices.
Currently, we cannot allocate > 25 clusters per allocation, because of
memory limits. The procedure requires the creation of a very large
matrix (2^n) that exceeds the available memory on most PCs (> 20GB). We
need to find a more efficient way of creating large matrices, perhaps
on-disk. This code is a start in that direction, using Python.